requirements.txt
opencv-python~=4.5.1.48

numpy~=1.20.0
matplotlib~=3.3.1
scikit-learn~=0.24.2
scikit-image~=0.18.1
six~=1.15.0
Pillow~=7.2.0
requests~=2.18.4
pandas~=1.2.0
lmdb~=1.2.1
natsort~=7.1.1

--find-links https://download.pytorch.org/whl/torch_stable.html
torch~=1.8.1+cu111
torchvision~=0.9.1+cu111

numba~=0.53.1

pyenchant~=3.2.0
Levenshtein~=0.12.0